changes in game design, analytics and revive logic

+++ Super Casual/Assets/Scripts/globals.cs
@@ -11,7 +11,7 @@ public class globals : MonoBehaviour {
     public int BALL_ID = 0;
     public float BALL_Y = -8f;
     public float BALL_X;
-    public float BALL_FLOOR = 0;
+    public int BALL_FLOOR = 0;
     public bool BALL_GROUNDED = true;
     public float BALL_SPEED_X = 4f;
     public float BALL_SPEED_Y;
